+++ src/ml/danneal.js
@@ -9,6 +9,6 @@
 
 // TODO
 // http://www.researchgate.net/profile/Joachim_Buhmann/publication/3192642_Pairwise_data_clustering_by_deterministic_annealing/links/0912f50c75945d835b000000.pdf
-FILTER.ML.deterministic_annealing = FILTER.ML.da = function(){};
+FILTER.MachineLearning.deterministic_annealing = /*FILTER.ML.da =*/ function(){};
 
 }(FILTER);